requirements.txt
streamlit
transformers
torch
graphviz
pylint
matplotlib
google-generativeai
PyPDF2
python-docx
pandas